+++ apps/ledger-live-mobile/src/components/RootNavigator/types/SignTransactionNavigator.ts
@@ -53,6 +53,7 @@ export type SignTransactionNavigatorParamList = {
     hideTotal?: boolean;
     appName?: string;
     dependencies?: string[];
+    isACRE?: boolean;
     currentNavigation:
       | ScreenName.SignTransactionSummary
       | ScreenName.SendSummary
@@ -71,6 +72,7 @@ export type SignTransactionNavigatorParamList = {
     status: TransactionStatus;
     appName?: string;
     dependencies?: string[];
+    isACRE?: boolean;
     onSuccess: (payload: unknown) => void;
     onError: (_: Error) => void;
     analyticsPropertyFlow?: string;