+++ libs/ledgerjs/packages/cryptoassets/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/cryptoassets
 
+## 13.6.1
+
+### Patch Changes
+
+- Updated dependencies [[`ced792c`](https://github.com/LedgerHQ/ledger-live/commit/ced792c37b42135f2b7596228c14ccd0783a803f)]:
+  - @ledgerhq/live-env@2.4.0
+
+## 13.6.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`ced792c`](https://github.com/LedgerHQ/ledger-live/commit/ced792c37b42135f2b7596228c14ccd0783a803f)]:
+  - @ledgerhq/live-env@2.4.0-next.0
+
 ## 13.6.0
 
 ### Minor Changes